細かい修正

+++ main.c
@@ -55,7 +55,7 @@ int main(int argc, char** argv)
 
   Vector* tokens = lex(fp);
 
-  for (int i = 0; i < tokens->length; i++) {
+  for (size_t i = 0; i < tokens->length; i++) {
     dump_token(*(Token*)(tokens->ptr[i]));
   }
   eprintf("\n");
@@ -64,7 +64,7 @@ int main(int argc, char** argv)
 
   Program* prog = parse(tokens);
 
-  for (int i = 0; i < prog->funcs->length; i++) {
+  for (size_t i = 0; i < prog->funcs->length; i++) {
     dump_function(prog->funcs->ptr[i]);
   }
